requirements.txt
python-dotenv >= 0.17.0
selenium >= 3.141.0
chromedriver-py == 92.0.4515.43